Add test to specific file
